+++ modules/tables/constants.js
@@ -143,6 +143,30 @@ export const HMCONST = {
         ENORMOUS: 7,
         COLOSSAL: 8,
     },
+    SKILL: {
+        DIFF: {
+            AUTO:       -1,
+            VDIFFICULT:  0,
+            DIFFICULT:   1,
+            AVERAGE:     2,
+            EASY:        3,
+            TRIVIAL:     4,
+        },
+        MASTERY: {
+            UNSKILLED:  0,
+            NOVICE:     1,
+            AVERAGE:    2,
+            ADVANCED:   3,
+            EXPERT:     4,
+            MASTER:     5,
+        },
+        TYPE: {
+            SKILL:      'value',
+            OPPOSED:    'opposed',
+            VERBAL:     'verbal',
+            WRITTEN:    'literacy',
+        },
+    },
     SPECIAL: {
         STANDARD:    0,
         JAB:         1,
@@ -521,11 +545,11 @@ export const HMTABLES = {
             turning:  'd20p +  @bonus.total.turning  + @resp.bonus',
             will:     'd20p +  @bonus.total.will     + @resp.bonus',
         },
-        'skill': {
-            'skill':    'd100 - (@resp.bonus + @bonus.total.value)',
-            'opposed':  'd100 + (@resp.bonus + @bonus.total.value)',
-            'language': 'd100 - (@resp.bonus + @bonus.total.verbal)',
-            'literacy': 'd100 - (@resp.bonus + @bonus.total.literacy)',
+        skill: {
+            [HMCONST.SKILL.TYPE.SKILL]:     'd100 - (@resp.bonus + @bonus.total.value)',
+            [HMCONST.SKILL.TYPE.OPPOSED]:   'd100 + (@resp.bonus + @bonus.total.value)',
+            [HMCONST.SKILL.TYPE.VERBAL]:    'd100 - (@resp.bonus + @bonus.total.verbal)',
+            [HMCONST.SKILL.TYPE.WRITTEN]:   'd100 - (@resp.bonus + @bonus.total.literacy)',
         },
     },
     bracket: {
@@ -603,14 +627,9 @@ export const HMTABLES = {
         [HMCONST.SCALE.ENORMOUS]: {hp: 35, kb: 40, reach:  5, token: 15, move:   6},
         [HMCONST.SCALE.COLOSSAL]: {hp: 70, kb: 75, reach: 12, token: 29, move:  13},
     },
-    'skill': {
-        'difficulty': {
-            'verydifficult': 10,
-            'difficult':      0,
-            'average':      -40,
-            'easy':         -80,
-            'trivial':      -90,
-        },
+    skill: {
+        difficulty: (arg) => [10, 0, -40, -80, -90].findIndex((m) => arg + m <= 0),
+        mastery: (arg) => [0, 25, 50, 75, 87, Infinity].findIndex((m) => m >= arg),
     },
     spell: {
         sfc: (svr) => {